fibonacciSequence.py
This program will print automatically the sequence for 0, 1, and 2 fibonacci sequence values.  Three or more values will be calculated and then returned.  When calculating 3 or more, since the initial list has two items already in it, the requested number of values from user input is reduced by two to calculate the proper amount of fibonacci numbers before printing.  In events of hanging code or unexpected errors, Keyboard Interrupt to exit.